Move to resource, check for VSCode
